Initial commit of Force Feedback autoceter test code
This is intended to clarify what happens under Windows vs Wine
with regard to the DIPROP_AUTOCENTER. In particular it aquires the
FF device with the default value, with DIPROP_AUTOCENTER_OFF, and
DIPROP_AUTOCENTER_ON, pausing each time to give time to play with
the device and allow the USB bus to be snoope with Wireshark.

The intent is to provide more insight for wine pull request 8605.

https://gitlab.winehq.org/wine/wine/-/merge_requests/8605